A personal project
This a combinations of header files created in C++. This helps to simplify arithmetic, relational operation. Instead of writing every operation just pass the values to the function and it will do the work itself. It can also find area and perimeter of a square and rectangle, area and circumference of a circle , odd/even, print no t to n.